Simplified hook
